fluentd-local-Elast-Kibana | swap up usecase46 to throw type errors

+++ UseCase46/src/V1/Worker/Delegate.php
@@ -9,10 +9,18 @@ class Delegate implements DelegateInterface
 {
     use Worker\Queue\Message\AwareTrait;
 
-    public function businessLogic(): DelegateInterface
+    public function businessLogic() : DelegateInterface
     {
-        $this->getV1WorkerQueueMessage()->delete();
+        if ((bool)random_int(0, 1)) {
+            $this->getV1WorkerQueueMessage()->delete();
+        } else {
+            $this->onlyTakesIntegers('Not very lucky');
+        }
 
         return $this;
+    }
+
+    public function onlyTakesIntegers(int $number){
+        return $number;
     }
 }
